matrix finalized, multi-class skimming working properly

+++ aequilibrae/distribution/gravity_application.py
@@ -185,7 +185,7 @@ class GravityApplication:
             a = self.columns.data[self.column_field][:]
 
             if self.model.function == "EXPO":
-                self.output.matrix[self.core_name][i, :] = np.exp(- self.model.beta * self.impedance.matrix_view[i, :, 0]) * p * a
+                self.output.matrix[self.core_name][i, :] = np.exp(- self.model.beta * self.impedance.matrix_view[i, :, 0].astype(np.float)) * p * a
 
             elif self.model.function == "POWER":
                 self.output.matrix[self.core_name][i, :] = np.nan_to_num(np.power(self.impedance.matrix_view[i, :, 0], - self.model.alpha) * p * a)[:]
